requirements.txt
pandas>=2.0
pyarrow>=12.0
scikit-learn>=1.3
lightgbm>=4.0
neo4j>=5.0
python-dotenv>=1.0
matplotlib>=3.7
seaborn>=0.12
numpy>=1.24
